--- selab_study_calculator/injun_calculator/Input.java
@@ -1,14 +1,0 @@
-package Calculate;
-
-import java.util.Scanner;
-
-public class Input {
-	
-	private static final Scanner scan = new Scanner(System.in);
-	
-	public static String input(){
-	System.out.println("ют╥б : ");
-	
-	return scan.nextLine();
-	}
-}
