--- selab_study_calculator/injun_calculator/Numbers.java
@@ -1,17 +1,0 @@
-package Calculate;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Numbers {
-	private List<Integer> numbers = new ArrayList<>();
-
-	protected void setNumbers(int number) {
-		this.numbers.add(number);
-	}
-	
-	protected List getNumbers() {
-		return numbers;
-	}
-	
-}
