--- selab_study_calculator/injun_calculator/Operators.java
@@ -1,16 +1,0 @@
-package Calculate;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Operators {
-	private List<String> opers = new ArrayList<>();
-	
-	protected void setOperators(String oper) {
-		this.opers.add(oper);
-	}
-	
-	protected List getOperators() {
-		return opers;
-	}
-}
