--- selab_study_calculator/injun_calculator/Output.java
@@ -1,9 +1,0 @@
-package Calculate;
-
-class Output {
-
-	protected static void print(String str){
-		System.out.println(str);
-	}
-	
-}
