+++ selab_study_calculator/src/Input.java
@@ -1,0 +1,12 @@
+package Calculate;
+
+import java.util.Scanner;
+
+public class Input {
+	private static final Scanner scan = new Scanner(System.in);
+
+	public static String input() {
+		System.out.println("Input : ");
+		return scan.nextLine();
+	}
+}
