+++ selab_study_calculator/src/Main.java
@@ -1,0 +1,10 @@
+package Calculate;
+	
+
+public class Main {
+	public static void main(String[] args) {
+		String word = Input.input();
+		new Separate(word);	
+	}
+
+}
