+++ selab_study_calculator/src/Numbers.java
@@ -1,0 +1,17 @@
+package Calculate;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Numbers {
+	private List<Integer> numbers = new ArrayList<>();
+
+	protected void setNumbers(int number) {
+		this.numbers.add(number);
+	}
+	
+	protected List<Integer> getNumbers() {
+		return numbers;
+	}
+	
+}
