+++ selab_study_calculator/src/Output.java
@@ -1,0 +1,8 @@
+package Calculate;
+
+class Output {
+	public static void print(String printword){
+		System.out.println(printword);
+	}
+	
+}
